Fixed Compile Time errors
