+++ src/main/java/com/company/functional/Main.java
@@ -5,5 +5,7 @@ import java.util.Map;
 public class Main {
 
     public static Map<String,Integer> getWordsWithOccurrenceInAscendingOrder(String inputString) {
+
+        return null;
     }
 }
