+++ src/main/java/com/company/maputils/MapUtils.java
@@ -31,20 +31,3 @@ public class MapUtils {
 }
 
 
-class MapAscendingSort implements Comparator {
-
-    private final Map map;
-
-    public MapAscendingSort(Map map) {
-        this.map = map;
-    }
-
-    public int compare(Object o1, Object o2) {
-        int comparison = ((Integer) map.get(o1)).intValue() - ((Integer) map.get(o2)).intValue();
-        if (comparison == 0) {
-            return ((String) o1).compareTo((String) o2);
-        } else {
-            return comparison;
-        }
-    }
-}
